+++ server/src/controllers/products/getProductsAllController.js
@@ -1,4 +1,4 @@
-const { Producto } = require("../db.js");
+const { Producto } = require("../../db.js");
 
 const getProductsAllController = async () =>{
   const data = await Producto.findAll();
