+++ server/src/routes/index.js
@@ -1,44 +1,23 @@
 const { Router } = require('express');
 const router = Router();
-
-// Importar todos los routers;
-// Ejemplo: const authRouter = require('./auth.js');
-
-// Router USUARIOS
 const usersRouter = require('./usersRouter');
 const productsRouter = require("./productsRouter")
 
-// Configurar los routers
-// Ejemplo: router.use('/auth', authRouter);
 
-// router Products
+// router PRODUCTOS
 router.use("/products", productsRouter)
 // Router USUARIOS
 router.use("/users", usersRouter);
-
-// router.get("/", startDatabase)
-
-router.post("/valoracion", postValoracion)
-router.put("/valoracion/:id", updateValoracion)
-router.delete("/valoracion/:id", deleteValoracion)
+// Router VALORACIONES
+// router.use("/valoracion", valoracionRouter);
 
 
-// Importar controladores
-const getAllProducts = require('../controllers/getAllProducts');
-const getProductById = require('../controllers/getProductById');
-const editProduct = require('../controllers/editProduct');
-
-
-// Definir rutas y controladores
-router.get("/productos", getAllProducts)
-router.get("/productos/:id", getProductById)
-router.patch("/productos/:id", editProduct);
-
-
-//valoracion
+//! Requiere cambios!
 const postValoracion = require('../controllers/postValoracion')
 const updateValoracion= require('../controllers/UpdateValoracion')
 const deleteValoracion= require('../controllers/deleteValoracion')
-
+router.post("/valoracion", postValoracion)
+router.put("/valoracion/:id", updateValoracion)
+router.delete("/valoracion/:id", deleteValoracion)
 
 module.exports = router;
